# Conflicts:
#	apl-utils/src/main/java/com/apollocurrency/aplwallet/apl/util/Constants.java

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/peer/DownloadableFilesManager.java
@@ -113,7 +113,7 @@ public class DownloadableFilesManager {
                 fci.fileId = fileId;
                 fci.offset = ci.offset;
                 fci.present = FileChunkState.PRESENT_IN_PEER;
-                fci.size = ci.size.longValue();
+                fci.size = ci.size;
                 fci.chunkId = i;
                 downloadInfo.chunks.add(fci);
             }
